feat: Major UI redesign + API fixes + in-app update feature
UI REDESIGN:
- Remove all TopAppBars for cleaner look
- Modern SkyBlue/Aqua color scheme throughout
- Floating pill-style bottom navigation with animations
- Premium card designs with gradients and shadows
- Beautiful section headers with accent bars
- Improved welcome, search, favorites, profile screens

API FIXES (CRITICAL):
- Fix AniList API URL: anilist.info → anilist.co
- Change score types: Int → Double for proper parsing
- Add comprehensive error logging
- Add network state checking before API calls
- Fix GraphQL error handling

NEW FEATURE - IN-APP UPDATER:
- GitHub Releases API integration
- Beautiful update dialog with download progress
- DownloadManager for reliable APK downloads
- Install flow after download completion
- Non-intrusive 3-second delay on startup

+++ app/src/main/java/com/anitrack/app/data/api/models/AnimeModel.kt
@@ -15,9 +15,9 @@ data class AnimeModel(
     @Json(name = "episodes")
     val episodes: Int? = null,
     @Json(name = "averageScore")
-    val averageScore: Int? = null,
+    val averageScore: Double? = null,
     @Json(name = "meanScore")
-    val meanScore: Int? = null,
+    val meanScore: Double? = null,
     @Json(name = "genres")
     val genres: List<String>? = null,
     @Json(name = "tags")
